requirements.txt
Flask==2.2.5
torch==2.0.1
torchvision==0.15.2
matplotlib==3.7.1
Pillow==9.5.0
numpy<2
